+++ src/codec_compare.ts
@@ -207,7 +207,7 @@ export class CodecCompare extends LitElement {
           </p>
 
           <p id="credits">
-            Codec Compare beta version 0.2.0<br>
+            Codec Compare beta version 0.2.1<br>
             <a href="https://github.com/webmproject/codec-compare">
               Sources on GitHub
             </a>